requirements.txt
requests
python-dotenv
aiogram
aiohttp==3.6.2
aioredis==1.3.1